Move options to discord/commands/options.py
